+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 - DataGrid: IsMoveAfterEnterEnabled property #158
 - PropertyGrid: support nullable enums #129
 - PropertyGrid: option to disable auto fill #143
+- DataGrid: register converters in cell definition factory #160
 
 ## [2.0.1]
 ### Changed